pythonProject/doc_meta_extractor/requirements.txt
Flask==3.0.0
python-docx==1.1.0
PyPDF2==3.0.1
Werkzeug==3.0.1